En Passant
En Passant is now functional (though the ruleset it runs off needs to be revised). Castling has been added but is not yet functional due to the rooks being set to can't castle before being moved. Also still need to add functionality to the Castling system to check if the cells between the king and chosen rook are empty.